+++ apps-common/pdf-templates/statistics-time-per-person.html
@@ -2,13 +2,13 @@
 <html lang="${ locale }">
 <head>
 	<meta charset="utf-8" />
-	<link href="${ root }statistics-time-per-role.html" rel="stylesheet" media="all" />
+	<link href="${ root }statistics-time-per-person.html" rel="stylesheet" media="all" />
 </head>
 <body>
 <header class="print-header">
 	${ logo ? '<img src="file://' + logo + '" />' : '' }
 	<div class="print-header-title">
-		<h2>${ e(_("Statistics time per role")) }</h2>
+		<h2>${ e(_("Statistics time per person")) }</h2>
 		<small>${ _("Printed on ${ currentDate }", { currentDate: currentDate }) }</small>
 	</div>
 </header>
