+++ routes/statistics.js
@@ -18,7 +18,10 @@ module.exports = function (processingStepsMeta) {
 			decorateContext: function () { this.processingStepsMeta = processingStepsMeta; },
 			view: require('../view/statistics-time-per-role')
 		},
-		'time/per-person': require('../view/statistics-time-per-person'),
+		'time/per-person': {
+			decorateContext: function () { this.processingStepsMeta = processingStepsMeta; },
+			view: require('../view/statistics-time-per-person')
+		},
 		analysis: require('../view/statistics-analysis'),
 
 		profile: require('../view/user-profile')
