+++ server/routes/statistics.js
@@ -1,16 +1,17 @@
 'use strict';
 
-var assign           = require('es5-ext/object/assign')
-  , normalizeOptions = require('es5-ext/object/normalize-options')
-  , ensureDriver     = require('dbjs-persistence/ensure-driver')
-  , ensureDatabase   = require('dbjs/valid-dbjs')
-  , ensureObject     = require('es5-ext/object/valid-object')
-  , QueryHandler     = require('../../utils/query-handler')
-  , getBaseRoutes    = require('./authenticated')
-  , timePerRolePrint = require('./statistics-time-per-role-print')
+var assign                            = require('es5-ext/object/assign')
+  , normalizeOptions                  = require('es5-ext/object/normalize-options')
+  , ensureDriver                      = require('dbjs-persistence/ensure-driver')
+  , ensureDatabase                    = require('dbjs/valid-dbjs')
+  , ensureObject                      = require('es5-ext/object/valid-object')
+  , QueryHandler                      = require('../../utils/query-handler')
+  , getBaseRoutes                     = require('./authenticated')
+  , timePerPersonPrint                = require('./statistics-time-per-person-print')
   , getProcessingTimesByStepProcessor =
 		require('../statistics/get-processing-times-by-step-processor')
-  , getQueryHandlerConf = require('../../routes/utils/get-statistics-time-query-handler-conf');
+  , getQueryHandlerConf               =
+		require('../../routes/utils/get-statistics-time-query-handler-conf');
 
 module.exports = exports = function (data) {
 	var options         = normalizeOptions(ensureObject(data)), queryConf, processingStepsMeta, db;
@@ -21,7 +22,7 @@ module.exports = exports = function (data) {
 		db: db,
 		processingStepsMeta: processingStepsMeta
 	});
-	timePerRolePrint = timePerRolePrint(assign(options));
+	timePerPersonPrint = timePerPersonPrint(assign(options));
 
 	var queryHandler = new QueryHandler(queryConf);
 
@@ -31,11 +32,11 @@ module.exports = exports = function (data) {
 				return getProcessingTimesByStepProcessor(assign(options, { query: query }));
 			});
 		},
-		'get-time-per-role-print': {
-			headers: timePerRolePrint.headers,
+		'get-time-per-person-print': {
+			headers: timePerPersonPrint.headers,
 			controller: function (query) {
 				return queryHandler.resolve(query)(function (query) {
-					return timePerRolePrint.controller({ query: query });
+					return timePerPersonPrint.controller({ query: query });
 				});
 			}
 		}
